Create build workflow
Chage Makefile to support PREFIX and DESTDIR
The build workflow will create 3 artifacts for different platform (armhf, arm64, amd64)
The version of the package is calculated based on latest release or pushed tag. As a fallback 0.0.1+<commit_sha> is used
If tags are pushed then a new release with that tag is created.

Signed-off-by: Nicu Siderias <nicu.siderias@analog.com>